Extract JSONValueMapper
Useful for other adapters which might want to use this, but not be
forced to require ActiveRecord.
