+++ lib/valkyrie/persistence.rb
@@ -31,6 +31,7 @@ module Valkyrie
     require 'valkyrie/persistence/composite_persister'
     require 'valkyrie/persistence/delete_tracking_buffer'
     require 'valkyrie/persistence/buffered_persister'
+    require 'valkyrie/persistence/shared'
     autoload :Postgres, 'valkyrie/persistence/postgres'
     autoload :Solr, 'valkyrie/persistence/solr'
     autoload :Fedora, 'valkyrie/persistence/fedora'
