+++ lib/valkyrie/persistence/shared.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+module Valkyrie::Persistence
+  module Shared
+    require 'valkyrie/persistence/shared/json_value_mapper'
+  end
+end
